Refactor migration services to enhance progress reporting and improve result properties, including updates to GroupPolicyMigrator and IdentityMigrator for better type handling and additional metadata.

+++ GUI/Services/Migration/GroupPolicyMigrator.cs
@@ -662,7 +662,7 @@ namespace MandADiscoverySuite.Services.Migration
                     // Report progress
                     var completedGpos = result.SuccessfulMigrations + result.FailedMigrations;
                     var progressPercentage = (double)completedGpos / result.TotalGpos * 100;
-                    context.ReportProgressUpdate("Bulk GPO Migration", progressPercentage,
+                    context.ReportProgressUpdate("Bulk GPO Migration", (int)progressPercentage,
                         $"Processed {completedGpos}/{result.TotalGpos} GPOs");
                 }
